+++ tests/testthat/helper-model.R
@@ -109,8 +109,8 @@ h_get_logistic_kadane <- function() {
 
 h_get_logistic_normal_mixture <- function() {
   LogisticNormalMixture(
-    comp1 = list(mean = c(0, 3), cov = diag(2)),
-    comp2 = list(mean = c(-1, 6), cov = c(2, 4) * diag(2)),
+    comp1 = ModelParamsNormal(mean = c(0, 3), cov = diag(2)),
+    comp2 = ModelParamsNormal(mean = c(-1, 6), cov = c(2, 4) * diag(2)),
     weightpar = c(a = 1, b = 5),
     ref_dose = 2
   )