+++ tests/testthat/helper-modelparams.R
@@ -1,0 +1,6 @@
+h_get_model_params_normal <- function() {
+  ModelParamsNormal(
+    mean = c(1, 6),
+    cov = diag(2)
+  )
+}